+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/implementation/converters/CreateChatThreadResultConverter.java
@@ -28,7 +28,7 @@ public final class CreateChatThreadResultConverter {
         }
 
         CreateChatThreadResult createChatThreadResult = new CreateChatThreadResult()
-            .setChatThread(ChatThreadConverter.convert(obj.getChatThread(), logger));
+            .setChatThreadProperties(ChatThreadPropertiesConverter.convert(obj.getChatThread(), logger));
 
         if (obj.getInvalidParticipants() != null) {
             List<ChatError> invalidParticipants